delete const isSeller, delete id owner from request, add check owner and throw exception

+++ src/main/java/ru/itis/master/party/dormdeals/services/impl/ShopsServiceImpl.java
@@ -4,6 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import ru.itis.master.party.dormdeals.dto.ProductDto.ProductDto;
 import ru.itis.master.party.dormdeals.dto.ProductDto.ProductsPage;
@@ -11,17 +13,24 @@ import ru.itis.master.party.dormdeals.dto.ShopDto.NewShop;
 import ru.itis.master.party.dormdeals.dto.ShopDto.ShopDto;
 import ru.itis.master.party.dormdeals.dto.ShopDto.ShopsPage;
 import ru.itis.master.party.dormdeals.dto.ShopDto.UpdateShop;
+import ru.itis.master.party.dormdeals.dto.UserDto.UserDto;
+import ru.itis.master.party.dormdeals.exceptions.NotAllowedException;
 import ru.itis.master.party.dormdeals.exceptions.NotFoundException;
 import ru.itis.master.party.dormdeals.models.Product;
+import ru.itis.master.party.dormdeals.models.Role;
 import ru.itis.master.party.dormdeals.models.Shop;
 import ru.itis.master.party.dormdeals.dto.ShopWithProducts;
 import ru.itis.master.party.dormdeals.models.User;
 import ru.itis.master.party.dormdeals.repositories.ProductsRepository;
 import ru.itis.master.party.dormdeals.repositories.ShopsRepository;
+import ru.itis.master.party.dormdeals.repositories.TokenRepository;
 import ru.itis.master.party.dormdeals.repositories.UserRepository;
 import ru.itis.master.party.dormdeals.services.ShopsService;
 
+import java.util.Objects;
+
 import static ru.itis.master.party.dormdeals.dto.ShopDto.ShopDto.from;
+import static ru.itis.master.party.dormdeals.dto.UserDto.UserDto.from;
 
 @Service
 @RequiredArgsConstructor
@@ -29,6 +38,10 @@ public class ShopsServiceImpl implements ShopsService {
     private final ShopsRepository shopsRepository;
     private final UserRepository userRepository;
     private final ProductsRepository productsRepository;
+
+    private final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    private User thisUser = userRepository.getByEmail(authentication.getName()).orElseThrow(() ->
+            new NotFoundException("Пользователь не найден"));
 
 
     @Value("${default.page-size}")
@@ -51,17 +64,16 @@ public class ShopsServiceImpl implements ShopsService {
     }
 
     @Override
-    public ShopDto createShop(NewShop newShop, Long ownerId) {
-        User owner = userRepository.findById(ownerId).orElseThrow(() -> new NotFoundException("Пользователь не найден"));
-
-        owner.setIsSeller(true);
-
+    public ShopDto createShop(NewShop newShop) {
+        thisUser.setRole(Role.ROLE_SELLER);
+//TODO: чтобы создание магазина отдавало дтошку, а не основного юзера
+//        UserDto ownerDto = from(user);
         Shop shop = Shop.builder()
                 .name(newShop.getName())
                 .description(newShop.getDescription())
                 .rating(newShop.getRating())
                 .placeSells(newShop.getPlaceSells())
-                .owner(owner)
+                .owner(thisUser)
                 .build();
 
         shopsRepository.save(shop);
@@ -71,13 +83,15 @@ public class ShopsServiceImpl implements ShopsService {
 
     @Override
     public ShopDto updateShop(Long id, UpdateShop updateShop) {
+
+        checkOwnerShop(id);
+
         Shop shopForUpdate = getShopOrThrow(id);
 
         shopForUpdate.setName(updateShop.getName());
         shopForUpdate.setDescription(updateShop.getDescription());
         shopForUpdate.setRating(updateShop.getRating());
         shopForUpdate.setPlaceSells(updateShop.getPlaceSells());
-//        shopForUpdate.setOwner(updateShop.getOwner());
 
         shopsRepository.save(shopForUpdate);
 
@@ -86,6 +100,9 @@ public class ShopsServiceImpl implements ShopsService {
 
     @Override
     public void deleteShop(Long id) {
+
+        checkOwnerShop(id);
+
         shopsRepository.deleteById(id);
     }
 
@@ -106,9 +123,13 @@ public class ShopsServiceImpl implements ShopsService {
                 .build();
     }
 
-    @Override
-    public Shop getShopOrThrow(long id) {
+    private Shop getShopOrThrow(long id) {
         return shopsRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Магазин с идентификатором <" + id + "> не найден"));
+    }
+    private void checkOwnerShop(Long shopId) {
+        if (!Objects.equals(thisUser.getId(), shopId)) {
+            throw new NotAllowedException("Вы не являетесь владельцем данного магазина.");
+        }
     }
 }
